+++ CHANGES.md
@@ -1,6 +1,6 @@
 2.2.0
 ---
-* ?
+* Rename have_enqueued_job to have_enqueued_sidekiq_job due to clash with rspec-rails [mrbrdo]
 
 2.1.0
 ---